Refactored the previous commit regarding wxIconBundle support for icons stored MS Windows resources.
