+++ include/wx/iconbndl.h
@@ -58,7 +58,7 @@ public:
     // initializes the bundle with a single icon
     wxIconBundle(const wxIcon& icon);
 
-#ifdef __WINDOWS__
+#if defined( __WINDOWS__) && wxUSE_ICO_CUR
     // initializes the bundle with the icons from a group icon stored as an MS Windows resource
     wxIconBundle(const wxString& resourceName, WXHINSTANCE module);
 #endif
@@ -75,7 +75,7 @@ public:
     void AddIcon(wxInputStream& stream, wxBitmapType type = wxBITMAP_TYPE_ANY);
 #endif // wxUSE_STREAMS && wxUSE_IMAGE
 
-#ifdef __WINDOWS__
+#if defined( __WINDOWS__) && wxUSE_ICO_CUR
     // loads all the icons from a group icon stored in an MS Windows resource
     void AddIcon(const wxString& resourceName, WXHINSTANCE module);
 #endif
